refactor: 리액션, 댓글 관련 API 로직 리팩토링 (#126)
